==> pipeline-v7/0.0_prepare-test-train.ipynb <==
import logging
logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

import pandas as pd
import numpy as np

df = pd.read_csv('../data/BBairline200722_coreffed.csv')
df.shape

df.sample(3, random_state=42)

# df = df.rename(
#     columns={
#         '0': 'url',
#         '1': 'date',
#         '2': 'title',
#         '3': 'authors',
#         '4': 'category',
#         '5': 'text'
#     }
# )
# df.sample(3, random_state=42)

from sklearn.model_selection import train_test_split

df_train, df_test = train_test_split(df, test_size=0.2, random_state=42)

df.shape, df_train.shape, df_test.shape

df.to_csv('./full.csv')

df_train.to_csv('./train.csv')

df_test.to_csv('./test.csv')
