==> pipeline-v7/2.1_lda-from-save.ipynb <==
import logging
logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

from gensim import corpora, models
import pyLDAvis.gensim_models

import simplejson

def json_load(filename):
    with open(filename, 'r', encoding='utf-8') as f:
        return simplejson.load(f)

dictionary = corpora.Dictionary.load_from_text('./lda.dictionary.txt')

corpus = corpora.MmCorpus('./lda.corpus.txt')

lda = models.ldamulticore.LdaMulticore.load('./lda.model')

lda.show_topics()

lda_vis = pyLDAvis.gensim_models.prepare(lda, corpus, dictionary, start_index=0, sort_topics=False)
pyLDAvis.display(lda_vis)

# dictionary_test = corpora.Dictionary.load_from_text('./lda.text.test.txt')

# corpus_test = corpora.MmCorpus('./lda.corpus.test.txt')

# corpus_words_test = json_load('./lda.corpus.test.json')

# import pandas as pd

# df_test = pd.read_csv('./test.csv')

# df_test.iloc[1].url

# TopicTags to each article ([1, 2, 3, 4...])
# Show articles of a specific topic tag
# TODO: Colour by topic

# TODO: Visualise articles in clusters - similarity on articles themselves

def docs2topics(corpus_words):
    return [lda[dictionary.doc2bow(doc)] for doc in corpus_words]

# #     for topic, sim in topics:
# #         print(lda.print_topic(topic))
# doc_topics_test = docs2topics(corpus_words_test)

#     for topic, sim in topics:
#         print(lda.print_topic(topic))
corpus_words_full = json_load('./lda.corpus.full.json')
doc_topics_full = docs2topics(corpus_words_full)

def articles_by_topic(df, doc_topics, topic_num, sim_thresh=0.5, attrs=["title", "url"]):
    for doc_id, topics in enumerate(doc_topics):
        for topic, sim in topics:
            if topic == topic_num:
                if sim > sim_thresh:
                    doc = df.iloc[doc_id]
                    print(f'{doc_id} {" ".join([doc[attr] for attr in attrs])} {sim}')

# articles_by_topic(df_test, doc_topics_test, 3)

import pandas as pd
df_full = pd.read_csv('./full.csv')

# def tag_articles()

articles_by_topic(df_full, doc_topics_full, 5)

# df_production = pd.read_csv('../data/BBairline200722.csv')

# corpus_words_production = json_load('./lda.corpus.production.json')

# doc_topics_production = docs2topics(corpus_words_production)

# articles_by_topic(df_production, doc_topics_production, 5, sim_thresh=0, attrs=["0", "2"])

from operator import itemgetter

doc_top_topics_full = [max(doc, key=itemgetter(1))[0] for doc in doc_topics_full]

doc_category_full = df_full.category.to_list()

import simplejson


def json_save(data, filename):
    with open(filename, 'w', encoding='utf-8') as f:
        simplejson.dump(data, f, separators=(',', ':'), iterable_as_array=True)

json_save(doc_category_full, './lda.doc_category_full.json')

json_save(doc_top_topics_full, './lda.doc_top_topics_full.json')

len(doc_top_topics_full)
